refactor: extract provider-agnostic map geometry helpers
