+++ app/(timeline)/fitness/heatmap/FitnessHeatmapView.test.tsx
@@ -20,12 +20,7 @@ import type {
 import { loadMapboxModule } from '@/lib/utils/mapbox'
 import { loadMaplibreModule } from '@/lib/utils/maplibre'
 
-import {
-  FitnessHeatmapView,
-  RouteHeatmapMap,
-  computeFocusBounds,
-  downsampleSegments
-} from './FitnessHeatmapView'
+import { FitnessHeatmapView, RouteHeatmapMap } from './FitnessHeatmapView'
 
 vi.mock('@/lib/utils/mapbox', () => ({
   loadMapboxModule: vi.fn()
@@ -1312,48 +1307,6 @@ describe('RouteHeatmapMap', () => {
     ).toBe(true)
   })
 
-  it('thins oversized geometry while preserving each segment’s endpoints', () => {
-    const longSegment = {
-      points: Array.from({ length: 12 }, (_, index) => ({
-        lat: index,
-        lng: index
-      }))
-    }
-    const shortSegment = {
-      points: [
-        { lat: 0, lng: 0 },
-        { lat: 1, lng: 1 }
-      ]
-    }
-
-    const result = downsampleSegments([longSegment, shortSegment], 6)
-    const [thinned, untouched] = result
-
-    // No segment is dropped — the route count is preserved.
-    expect(result).toHaveLength(2)
-    expect(thinned.points.length).toBeLessThan(longSegment.points.length)
-    expect(thinned.points[0]).toEqual(longSegment.points[0])
-    expect(thinned.points[thinned.points.length - 1]).toEqual(
-      longSegment.points[longSegment.points.length - 1]
-    )
-    // Segments with two or fewer points are left exactly as-is.
-    expect(untouched).toBe(shortSegment)
-  })
-
-  it('returns the original segments unchanged when under the budget', () => {
-    const segments = [
-      {
-        points: [
-          { lat: 0, lng: 0 },
-          { lat: 1, lng: 1 },
-          { lat: 2, lng: 2 }
-        ]
-      }
-    ]
-
-    expect(downsampleSegments(segments, 100)).toBe(segments)
-  })
-
   it('fits the full bounds (no zoom cap) for a single contiguous region', async () => {
     const { Map, fitBounds } = createFitBoundsCapturingModule()
     mockLoadMaplibreModule.mockResolvedValue({ Map })
@@ -1390,260 +1343,6 @@ describe('RouteHeatmapMap', () => {
       padding: 56,
       duration: 0,
       maxZoom: 12
-    })
-  })
-})
-
-describe('computeFocusBounds', () => {
-  it('keeps the full bounds for a single contiguous region', () => {
-    const bounds = { minLat: 52.36, maxLat: 52.39, minLng: 4.88, maxLng: 4.91 }
-    const result = computeFocusBounds(
-      [
-        {
-          points: [
-            { lat: 52.36, lng: 4.88 },
-            { lat: 52.39, lng: 4.91 }
-          ]
-        }
-      ],
-      bounds
-    )
-
-    expect(result.focused).toBe(false)
-    expect(result.bounds).toBe(bounds)
-  })
-
-  it('keeps the full bounds for a region spanning several adjacent cells', () => {
-    // Points spread contiguously across ~8° of longitude (several 5° cells that
-    // are 8-connected), so there is a single cluster — show the whole extent.
-    const bounds = { minLat: 50, maxLat: 52, minLng: 4, maxLng: 12 }
-    const result = computeFocusBounds(
-      [
-        {
-          points: [
-            { lat: 50, lng: 4 },
-            { lat: 51, lng: 8 },
-            { lat: 52, lng: 12 }
-          ]
-        }
-      ],
-      bounds
-    )
-
-    expect(result.focused).toBe(false)
-    expect(result.bounds).toBe(bounds)
-  })
-
-  it('tightens to the densest cluster for disjoint regions', () => {
-    const bounds = { minLat: 1.3, maxLat: 52.39, minLng: 4.88, maxLng: 103.9 }
-    const result = computeFocusBounds(
-      [
-        {
-          points: [
-            { lat: 52.36, lng: 4.88 },
-            { lat: 52.39, lng: 4.91 }
-          ]
-        },
-        { points: SINGAPORE_CLUSTER_POINTS }
-      ],
-      bounds
-    )
-
-    expect(result.focused).toBe(true)
-    expect(result.bounds).toEqual(SINGAPORE_CLUSTER_BOUNDS)
-  })
-
-  it('ignores a sparse far-away outlier and frames the main cluster', () => {
-    const bounds = { minLat: 52.36, maxLat: 52.39, minLng: 4.88, maxLng: 40 }
-    const result = computeFocusBounds(
-      [
-        {
-          points: [
-            { lat: 52.36, lng: 4.88 },
-            { lat: 52.37, lng: 4.89 },
-            { lat: 52.39, lng: 4.91 }
-          ]
-        },
-        // A single stray point far east in its own, disconnected cell.
-        { points: [{ lat: 52.36, lng: 40 }] }
-      ],
-      bounds
-    )
-
-    expect(result.focused).toBe(true)
-    expect(result.bounds).toEqual({
-      minLat: 52.36,
-      maxLat: 52.39,
-      minLng: 4.88,
-      maxLng: 4.91
-    })
-  })
-
-  it('returns the full bounds for an empty segment list', () => {
-    const bounds = { minLat: 0, maxLat: 0, minLng: 0, maxLng: 0 }
-    const result = computeFocusBounds([], bounds)
-
-    expect(result.focused).toBe(false)
-    expect(result.bounds).toBe(bounds)
-  })
-
-  it('skips non-finite vertices so they create no spurious cluster', () => {
-    const bounds = { minLat: 52.36, maxLat: 52.39, minLng: 4.88, maxLng: 4.91 }
-    const result = computeFocusBounds(
-      [
-        {
-          points: [
-            { lat: 52.36, lng: 4.88 },
-            { lat: 52.39, lng: 4.91 }
-          ]
-        },
-        // An entirely non-finite segment must contribute no grid cell; otherwise
-        // it would look like a second region and flip the result to focused.
-        {
-          points: [
-            { lat: NaN, lng: NaN },
-            { lat: Infinity, lng: -Infinity }
-          ]
-        }
-      ],
-      bounds
-    )
-
-    expect(result.focused).toBe(false)
-    expect(result.bounds).toBe(bounds)
-  })
-
-  it('ignores non-finite vertices when framing a focused cluster', () => {
-    const bounds = { minLat: 1.3, maxLat: 52.39, minLng: 4.88, maxLng: 103.9 }
-    const result = computeFocusBounds(
-      [
-        {
-          points: [
-            { lat: 52.36, lng: 4.88 },
-            { lat: 52.37, lng: 4.89 },
-            { lat: 52.39, lng: 4.91 }
-          ]
-        },
-        // The denser cluster carries a stray non-finite vertex that must not
-        // widen (or NaN-poison) the focused box.
-        { points: [{ lat: NaN, lng: 103.8 }, ...SINGAPORE_CLUSTER_POINTS] }
-      ],
-      bounds
-    )
-
-    expect(result.focused).toBe(true)
-    expect(result.bounds).toEqual(SINGAPORE_CLUSTER_BOUNDS)
-  })
-
-  it('does not merge clusters straddling the antimeridian (documented limitation)', () => {
-    // Two clusters at opposite signs near ±180° lon fall in non-adjacent grid
-    // cells, so the focus frames only the denser one rather than spanning the
-    // shorter way around the globe.
-    const bounds = { minLat: 0, maxLat: 1, minLng: -179.9, maxLng: 179.9 }
-    const result = computeFocusBounds(
-      [
-        {
-          points: [
-            { lat: 0.5, lng: 179.5 },
-            { lat: 0.6, lng: 179.7 },
-            { lat: 0.55, lng: 179.9 }
-          ]
-        },
-        { points: [{ lat: 0.5, lng: -179.9 }] }
-      ],
-      bounds
-    )
-
-    expect(result.focused).toBe(true)
-    expect(result.bounds).toEqual({
-      minLat: 0.5,
-      maxLat: 0.6,
-      minLng: 179.5,
-      maxLng: 179.9
-    })
-  })
-
-  it('merges cells that touch only diagonally (8-connectivity)', () => {
-    const bounds = { minLat: 52, maxLat: 55, minLng: 4, maxLng: 8 }
-    // The two 5° cells (0:10 and 1:11) are diagonal neighbours — connected only
-    // at a corner. 8-connectivity merges them into one contiguous cluster, so
-    // the full bounds are kept (a 4-connected flood fill would NOT merge these).
-    const result = computeFocusBounds(
-      [
-        {
-          points: [
-            { lat: 52, lng: 4 },
-            { lat: 55, lng: 8 }
-          ]
-        }
-      ],
-      bounds
-    )
-
-    expect(result.focused).toBe(false)
-    expect(result.bounds).toBe(bounds)
-  })
-
-  it('does not merge cells a knight’s-move apart, focusing the densest cell', () => {
-    const bounds = { minLat: 52, maxLat: 55, minLng: 4, maxLng: 13 }
-    // Cell 0:10 (two points) is strictly denser than cell 2:11 (one point), and
-    // the two cells are a knight's move apart (dx=2), so they stay separate — this
-    // pins that grid adjacency (not mere proximity) is what merges clusters, and
-    // that the seed is chosen by density rather than Map insertion order.
-    const result = computeFocusBounds(
-      [
-        {
-          points: [
-            { lat: 52, lng: 4 },
-            { lat: 53, lng: 4 },
-            { lat: 55, lng: 13 }
-          ]
-        }
-      ],
-      bounds
-    )
-
-    expect(result.focused).toBe(true)
-    expect(result.bounds).toEqual({
-      minLat: 52,
-      maxLat: 53,
-      minLng: 4,
-      maxLng: 4
-    })
-  })
-
-  it('frames the densest cell even when another cluster has more cells', () => {
-    // Cluster A spans two 8-adjacent cells (0:0 and 0:1) of one point each — the
-    // largest connected cluster by cell count. Cluster B is a single isolated cell
-    // (8:0) holding three points — the densest cell. The focus must frame B (seed
-    // = densest cell, then its cluster), proving the helper does not instead pick
-    // the cluster with the most cells.
-    const bounds = { minLat: 0, maxLat: 7, minLng: 2, maxLng: 40.2 }
-    const result = computeFocusBounds(
-      [
-        {
-          points: [
-            { lat: 2, lng: 2 },
-            { lat: 7, lng: 2 }
-          ]
-        },
-        {
-          points: [
-            { lat: 0, lng: 40 },
-            { lat: 0.1, lng: 40.1 },
-            { lat: 0.2, lng: 40.2 }
-          ]
-        }
-      ],
-      bounds
-    )
-
-    expect(result.focused).toBe(true)
-    expect(result.bounds).toEqual({
-      minLat: 0,
-      maxLat: 0.2,
-      minLng: 40,
-      maxLng: 40.2
     })
   })
 })
